HBX-1992: HBX-1992: Reorganize the reverse engineering strategy hierarchy - Rename 'org.hibernate.tool.internal.reveng.DelegatingReverseEngineeringStrategy' to 'org.hibernate.tool.internal.reveng.strategy.DelegatingStrategy'
